Moved the Option extension into the option class, no reason to have an extension when we control the class
