+++ Mittons.Fixtures.Tests.Unit/Extensions/OptionExtensionsTests.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
 using Mittons.Fixtures.Extensions;
 using Mittons.Fixtures.Models;
@@ -10,87 +9,13 @@ namespace Mittons.Fixtures.Tests.Unit.Extensions;
 public class OptionExtensionsTests
 {
     [Fact]
-    public void ToExecutionParameterFormattedString_WhenOptionIsNull_ExpectArgumentNullException()
-    {
-        Option? stubOption = null;
-
-        Assert.Throws<ArgumentNullException>(() => stubOption.ToExecutionParameterFormattedString());
-    }
-
-    [Fact]
-    public void ToExecutionParameterFormattedString_WhenNameIsNull_ExpectArgumentNullException()
-    {
-        var stubOption = new Option
-        {
-            Name = null,
-            Value = "value"
-        };
-
-        Assert.Throws<ArgumentNullException>(() => stubOption.ToExecutionParameterFormattedString());
-    }
-
-    [Fact]
-    public void ToExecutionParameterFormattedString_WhenValueIsNull_ExpectExecutionParameterFormattedStringWithOnlyName()
-    {
-        var stubOption = new Option
-        {
-            Name = "--option",
-            Value = null
-        };
-
-        var commandFormattedString = stubOption.ToExecutionParameterFormattedString();
-
-        Assert.Equal("--option", commandFormattedString);
-    }
-
-    [Fact]
-    public void ToExecutionParameterFormattedString_WhenValueIsEmptyString_ExpectExecutionParameterFormattedStringWithOnlyName()
-    {
-        var stubOption = new Option
-        {
-            Name = "--option",
-            Value = string.Empty
-        };
-
-        var commandFormattedString = stubOption.ToExecutionParameterFormattedString();
-
-        Assert.Equal("--option", commandFormattedString);
-    }
-
-    [Fact]
-    public void ToExecutionParameterFormattedString_WhenValueIsWhiteSpace_ExpectExecutionParameterFormattedStringWithOnlyName()
-    {
-        var stubOption = new Option
-        {
-            Name = "--option",
-            Value = " "
-        };
-
-        var commandFormattedString = stubOption.ToExecutionParameterFormattedString();
-
-        Assert.Equal("--option", commandFormattedString);
-    }
-
-    [Fact]
-    public void ToExecutionParameterFormattedString_WhenNameAndValueAreNotNull_ExpectExecutionParameterFormattedString()
-    {
-        var stubOption = new Option
-        {
-            Name = "--option",
-            Value = "value"
-        };
-
-        var commandFormattedString = stubOption.ToExecutionParameterFormattedString();
-
-        Assert.Equal("--option \"value\"", commandFormattedString);
-    }
-
-    [Fact]
-    public void ToExecutionParametersFormattedString_WhenOptionsIsNull_ExpectArgumentNullException()
+    public void ToExecutionParametersFormattedString_WhenOptionsIsNull_ExpectEmptyString()
     {
         IEnumerable<Option>? stubOptions = null;
 
-        Assert.Throws<ArgumentNullException>(() => stubOptions.ToExecutionParametersFormattedString());
+        var commandFormattedString = stubOptions.ToExecutionParametersFormattedString();
+
+        Assert.Equal(string.Empty, commandFormattedString);
     }
 
     [Fact]
