+++ Mittons.Fixtures/Models/Option.cs
@@ -1,4 +1,6 @@
-﻿namespace Mittons.Fixtures.Models
+﻿using System;
+
+namespace Mittons.Fixtures.Models
 {
     /// <summary>
     /// A docker option used to further configure a command.
@@ -14,9 +16,25 @@
         /// The value for the option.
         /// </summary>
         /// <remarks>
-        /// Set as a direct value with no quotations.  
+        /// Set as a direct value with no quotations.
         /// Set to <c>null</c>, <c>empty string</c>, or <c>white-space</c> if there is no value to accompany the option name.
         /// </remarks>
         public string Value { get; set; }
+
+        /// <summary>
+        /// The parameter string that can be inserted in the execution string.
+        /// </summary>
+        public string ExecutionParameter
+        {
+            get
+            {
+                if (Name is null)
+                {
+                    throw new ArgumentNullException();
+                }
+
+                return string.IsNullOrWhiteSpace(Value) ? Name : $"{Name} \"{Value}\"";
+            }
+        }
     }
 }
